requirements.txt
flask==3.1.0
python-docx==1.1.2
Pillow==11.1.0
gunicorn